+++ btrfs_timeline/__init__.py
@@ -2,4 +2,4 @@
 # -*- coding: utf-8 -*-
 """btrfs-timeline: btrfs スナップショットからファイルの過去版を辿り、復元する。"""
 
-__version__ = '0.4.0'
+__version__ = '0.5.0'